Native 1.2: moved all 7.* section types to native.*
