+++ native/ad_unit.go
@@ -1,12 +1,12 @@
-package request
+package native
 
-// 7.2 Native Ad Unit IDs - To Be Deprecated
-//
-// Ad Unit ID is to be deprecated in a future version and is not suggested for new implementations.
+// 7.2 Native Ad Unit IDs in version 1.1.
+// Deprecated/removed in version 1.2.
 //
 // Below is a list of the core ad unit ids described by IAB here http://www.iab.net/media/file/IABNativeAdvertisingPlaybook120413.pdf
 //
-// In feed unit is essentially a layout, it has been removed from the list. The in feed units can be identified via the layout parameter on the request.
+// In feed unit is essentially a layout, it has been removed from the list.
+// The in feed units can be identified via the layout parameter on the request.
 //
 // An implementing exchange may not support all asset variants or introduce new ones unique to that system.
 type AdUnit int64
